musicgen: Remove CL cache manager
As GPU devices now use stateful models, we don't need to explicitly manage KV cache.

This allows us to remove the CL cache manager, as it's no longer used.

Furthermore, as this was the only direct use of OpenCL for this project,
we also remove OpenCL as a build / runtime dependency.